css/css-borders-4/parsing/box-shadow-spread-invalid.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>CSS Borders 4 Test: Parsing 'box-shadow-spread' with invalid values</title>
        <link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com" />
        <link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-spread" />
        <meta name="assert" content="'box-shadow-spread' supports only properly defined length values" />
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script src="/css/support/parsing-testcommon.js"></script>
	</head>
	<body>
		<script>
		test_invalid_value("box-shadow-spread", "auto");
		test_invalid_value("box-shadow-spread", "none");
		test_invalid_value("box-shadow-spread", "1");
		test_invalid_value("box-shadow-spread", "1px 2px");
		test_invalid_value("box-shadow-spread", "1%");
		</script>
	</body>
</html>
